requirements.txt
# Core dependencies
numpy==1.26.4
pandas==2.2.2
scikit-learn==1.4.2
scipy==1.13.1
joblib==1.4.2
tqdm==4.66.4
matplotlib==3.8.4
seaborn==0.13.2

# AWS and cloud
boto3==1.34.102
botocore==1.34.102
s3transfer==0.10.1

# FastAPI and serving
uvicorn==0.30.1
click==8.1.7
fastapi==0.104.0

# ML preprocessing
pyarrow==15.0.2
protobuf==4.25.3
threadpoolctl==3.5.0


# Data science utils
python-dateutil==2.9.0.post0
pytz==2024.1
text-unidecode==1.3
python-slugify==8.0.4
packaging==24.0

# Dev/debug
ipython==8.12.3
jedi==0.19.1
parso==0.8.3
prompt-toolkit==3.0.43
pygments==2.18.0
traitlets==5.14.2


# Kaggle and data utils
kaggle==1.6.12
requests==2.31.0
certifi==2024.2.2
charset-normalizer==3.3.2
idna==3.7
urllib3==1.26.18

# Machine learning and embeddings
faiss-cpu==1.7.4

# frontend
streamlit
requests